# openapi.yaml
openapi: 3.0.1
info:
  title: YFinance API Backend
  version: 1.1.0
  description: |
    A simple API wrapper around the Python yfinance library.
    Includes endpoints for price, history, news, and sentiment analysis.

paths:

  /:
    get:
      summary: Health check
      responses:
        '200':
          description: API status

  /price/{symbol}:
    get:
      summary: Get current price for a ticker
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: AAPL
      responses:
        '200':
          description: Price information
          content:
            application/json:
              schema:
                type: object

  /history/{symbol}:
    get:
      summary: Get historical OHLC data
      parameters:
        - name: symbol
          in: path
          required: true
        - name: period
          in: query
          schema:
            type: string
            default: 1mo
        - name: interval
          in: query
          schema:
            type: string
            default: 1d
      responses:
        '200':
          description: Historical data

  /news/{symbol}:
    get:
      summary: Get news for a ticker
      parameters:
        - name: symbol
          in: path
          required: true
      responses:
        '200':
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  /sentiment/{symbol}:
    get:
      summary: Sentiment analysis for ticker news
      description: Returns VADER sentiment scores for each news article and an aggregated score.
      parameters:
        - name: symbol
          in: path
          required: true
      responses:
        '200':
          description: Sentiment scores
          content:
            application/json:
              schema:
                type: object